base changes to split CompetitionSpecification into prediction and model based
